requirements.txt
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1635819655903/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
certifi==2021.10.8
click==8.0.3
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1636043384856/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
gensim==4.1.2
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1638555527013/work/dist/ipykernel-6.6.0-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1638470588570/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1637175406194/work
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1636814410433/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
nltk==3.6.5
nose==1.3.7
numpy==1.21.4
pandas==1.3.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
python-crfsuite==0.9.7
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1635877545254/work
regex==2021.11.10
scikit-learn==1.0.1
scipy==1.7.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn-crfsuite==0.3.6
smart-open==5.2.1
tabulate==0.8.9
threadpoolctl==3.0.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1635819729986/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work